Move iframe utils into shared
