+++ frontend/src/metabase/dashboard/actions/utils.ts
@@ -1,9 +1,9 @@
 import { t } from "ttag";
 import _ from "underscore";
 
-import { getIframeDomainName } from "metabase/dashboard/visualizations/IFrameViz/utils";
 import type { StoreDashboard, StoreDashcard } from "metabase/redux/store";
 import { isActionDashCard } from "metabase/utils/dashboard";
+import { getIframeDomainName } from "metabase/visualizations/lib/iframe";
 import type {
   DashCardId,
   Dashboard,
