+++ frontend/src/metabase/dashboard/visualizations/IFrameViz/IFrameViz.tsx
@@ -9,6 +9,10 @@ import { getParameterValues } from "metabase/dashboard/selectors";
 import { useSelector } from "metabase/redux";
 import { getUserIsAdmin } from "metabase/selectors/user";
 import { Box, Button, Group, Icon, Stack, Text } from "metabase/ui";
+import {
+  getAllowedIframeAttributes,
+  isAllowedIframeUrl,
+} from "metabase/visualizations/lib/iframe";
 import { fillParametersInText } from "metabase/visualizations/shared/utils/parameter-substitution";
 import type {
   Dashboard,
@@ -23,7 +27,6 @@ import {
   StyledInput,
 } from "./IFrameViz.styled";
 import { settings } from "./IFrameVizSettings";
-import { getAllowedIframeAttributes, isAllowedIframeUrl } from "./utils";
 
 export interface IFrameVizProps {
   dashcard: VirtualDashboardCard;
