+++ frontend/src/metabase/parameters/utils/mapping-options.ts
@@ -3,7 +3,6 @@ import _ from "underscore";
 
 import { tag_names } from "cljs/metabase.parameters.shared";
 import { getColumnIcon } from "metabase/common/utils/columns";
-import { getAllowedIframeAttributes } from "metabase/dashboard/visualizations/IFrameViz/utils";
 import { getGroupName } from "metabase/querying/filters/utils/groups";
 import type { IconName } from "metabase/ui";
 import {
@@ -11,6 +10,7 @@ import {
   isQuestionDashCard,
   isVirtualDashCard,
 } from "metabase/utils/dashboard";
+import { getAllowedIframeAttributes } from "metabase/visualizations/lib/iframe";
 import * as Lib from "metabase-lib";
 import { TemplateTagDimension } from "metabase-lib/v1/Dimension";
 import type { DimensionOptionsSection } from "metabase-lib/v1/DimensionOptions/types";
